Workspace Package: * Create __init__ * Move workspace_reader to Workspace Package * Move xml_reader to Workspace Package
